Addition of an automated taask to generate the provider docs
Fixes: #94

the docs are generated using github.com/hashicorp/terraform-plugin-docs

Any resource/datasource will be automatically scaffolded with a new docs file.
To add an example to that scaffolding, the examples folder now has
resource / provider and data-sources folders

Inside each of these fodlers are the appropriate resources and an example
and this example is ingested as part of the docs generation

I had to move docs/github around as the docs generator removes these. I
moved the ISSUE and PULL_REQUEST templates to the .github folder so they
can be used and the others to the root

This provider will be used as the scaffolding for the upstream
[Pulumi](https://www.pulumi.com/) provider so it was in our interests to
make sure that this has documentation for our users