Add AGPL 3.0 License to satisfy madmin
Since v2 of madmin (the SDK for administrative access to minio)
switched to AGPL 3, the licenses used in this project are:

      1 AGPL-3.0
     18 Apache-2.0
      3 BSD-2-Clause
     14 BSD-3-Clause
     20 MIT
     14 MPL-2.0

Since AGPL 3 is a copyleft license, as opposed to the more permissive
licenses before, this has to be addressed.  At least the combined
work (on building the provider) and publishing it to the terraform
registry I believe must be done under AGPL 3.0.

* see https://min.io/compliance
* see https://www.gnu.org/licenses/license-list.html#GPLCompatibleLicenses
* see https://en.wikipedia.org/wiki/GNU_General_Public_License#Compatibility_and_multi-licensing
* see https://softwareengineering.stackexchange.com/a/371435