Rename `bandwidth_limit` back to `bandwidth_limt` (#630)
This PR is the first part of resolution of the issue #629 and introducing the following change:
renaming `bandwidth_limit` back to `bandwidth_limt` which was incorrectly renamed in #623.

Renaming the optional field leads to an inability to plan a terraform configuration if the `minio_s3_bucket_replication` resource was created in versions prior to `3.4.0` with provider versions later than `3.4.0`.

The plan to fix this issue consists of two steps:
1. Rename `bandwidth_limit` back to `bandwidth_limt` and release as version `3.5.1`
2. Mark `bandwidth_limt` as obsolete, add the new `bandwidth_limit` attribute and implement migration logic and release as version `3.5.2`
See comments in the issue: https://github.com/aminueza/terraform-provider-minio/issues/629#issuecomment-2862509136

This pull request implements the first step of the fix.

Merging this PR allows users to safely upgrade this provider from versions prior to `3.4.0` to version `3.5.1` and then upgrade to version `3.5.2` when it is released.