Consider the policy attachment as removed if user is removed (#635)
## Main change

Resolves #614

Since the policy attachment in the MinIO API is an attribute of the user, we consider that if the user is removed, the policy attachment is removed as well.

## Additional changes

Added the `main.tf` and `variables.tf` to the `iam_user_policy_attachment` example to make it runnable with docker-compose stack from the repo.